requirements.txt
djangorestframework==3.15.2
drf-spectacular==0.28.0
